+++ bluefin/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add `runPureEffAsyncSafe` and `runPureEffAsyncSafeRestarting`
+
 * Use `INLINE [0]` on `runDslBuilderEff` for improved performance
 
 # 0.10.0.0